+++ tests/common/mod.rs
@@ -1,5 +1,0 @@
-use std::fs;
-
-pub fn load_test_file(file: &str) -> String {
-    fs::read_to_string(file).unwrap_or_else(|_| panic!("Failed to read test file: {file}",))
-}